+++ src/pbrt/integrators/ambient_occlusion.h
@@ -9,7 +9,7 @@
 
 class AmbientOcclusionIntegrator {
   public:
-    void init(const Spectrum *_illuminant_spectrum, const double _illuminant_scale) {
+    void init(const Spectrum *_illuminant_spectrum, const FloatType _illuminant_scale) {
         illuminant_spectrum = _illuminant_spectrum;
         illuminant_scale = _illuminant_scale;
     }
@@ -51,5 +51,5 @@ class AmbientOcclusionIntegrator {
 
   private:
     const Spectrum *illuminant_spectrum;
-    double illuminant_scale;
+    FloatType illuminant_scale;
 };
